Added visualisation module for simulation data rendering
- Implemented a new module `Visualisation.py` to handle the rendering of simulation data.
- Added functions to create and save animations in MP4 and GIF formats using Matplotlib.
- Included functionality for interactive frame selection via a slider.
- Integrated various plots for speed, pressure, Mach number, density, and forces over time.
- Ensured compatibility with compressible and incompressible flow simulations.
- Added colorbars and titles for clarity in visual representation.
